fix: fix double filters
